// src/app/OT_DSR/catalogo/servicios/interfaces/servicio.interface.ts
// Generated by https://quicktype.io

export interface ServiciosResponse {
  id:     number;
  codigo: string;
  nombre: string;
  estado: boolean;
  fecha:  string;
  precio: number;
}

export interface ServiciosForm {
  id?:     number;
  codigo: string;
  nombre: string;
  estado?: boolean;
  fecha?:  string;
  precio: number;
}